% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read fars}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{(chr) Complete path to the fars file to read.}
}
\value{
A \code{data_frame}.
}
\description{
Read a fars file  returning the imported \code{data_frame}. An error is
returned if the path is not valid.
}
\details{
Fars file has to be stored in a \code{CSV} file format.
  It can be possibly compressed (see \code{\link{file}}).
}
\examples{
\dontrun{
    fars_read("accident_2013.csv.bz2")
}
}
