% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read multiple fars}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{(num) An integer vector of years (or an object
cohercible to it) containing the years of fars files
of interest.}
}
\value{
A list of \code{data_frame}, one for each year privided in
        input, with two columns: \code{MONTH}, and \code{year} of the
        corresponding fars file. If the file does not exist
        \code{NULL} is stored in the corresponding slot of the list.
}
\description{
Import a set of fars files given the vector of their years.
}
\details{
Computation does not stop if the file does not exist: a
   warning is produced and \code{NULL} is stored in the corresponding
   slot of the output list.
}
\examples{
\dontrun{
    fars_read_years(c(2012, 2013, 2014))                # one warning
    fars_read_years(c(2013, 2014))                       # no warning
    fars_read_years(c("2012", "2013", "2014"))# works with characters
}
}
