% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_filename.R
\name{make_filename}
\alias{make_filename}
\title{Make fars filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{(num) The year of the interested fars file.}
}
\value{
A character string of length one with the filename
    (full extension included) of the fars file for the year of
    interest.
}
\description{
Compose the correct name for a fars file given the reference year.
}
\note{
This function is for internal usage only and it is not
      exported.
}
\examples{
\dontrun{
    weektwo:::make_filename("2012")
}
}
