% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Map accidents}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{(num) An integer (or something cohercible to it)
representing a STATE number into the fars data.}

\item{year}{(num) The year of the interested fars file.}
}
\value{
\code{NULL} (with a message) if the STATE represented by the
        \code{state.num} provided has no records in fars for the year
        selected. Otherwhise, a plot with the geographical point for
        its fars is produced and nothing will be returned.
}
\description{
Plot the point map for the fars record of the STATE corresponding to
the \code{state.num} and the \code{year} provided
}
\details{
if no accident happen in the state and period selected
         a message will be throw and \code{NULL} is returned.

         If a plot is returned, it will be drawn only for longitude
         not greater than 900 and latidude not grater than 90.
}
\examples{
\dontrun{
    fars_map_state(1, 2013)
    fars_map_state(1, 2012) # error
    fars_map_state(2, 2015) # nothing to drawn
}
}
