% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_data.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Draw a map for a given state number and a given year.}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{a numeric value to indicate the state number.}

\item{years}{a numeric value of a specific year.}
}
\value{
a map satisfied the condition.
}
\description{
This function is used to draw a map for a specific state number given a specific year.
External packages are loaded including dplyr (for filtering a subset of the data)
and maps, graphics (for drawing a plot).
}
\examples{
fars_map_state(1,2013)

}
