% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_read.R
\name{fars_read}
\alias{fars_read}
\title{Read fars files}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{(chr) Complete path to the fars file to read.}

\item{path}{(chr) path to the fars file}
}
\value{
A \code{\link[dplyr]{data_frame}}.
}
\description{
Read a fars file returning the imported
\code{\link[dplyr]{data_frame}}.
}
\details{
An error is returned if the path is not valid.

Fars file has to be stored in a \code{CSV} file format.
  It can be possibly compressed (see \code{\link{file}}).
}
\examples{
\dontrun{
   file_path <- system.file("tests", "testthat", package = "weektwo",
       "accident_2013.csv.bz2"
   )
   fars_read(file_path)
}
}
