% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_data.R
\name{fars_read}
\alias{fars_read}
\title{Read a csv file}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{The name of a csv file.}
}
\value{
A tibble data framework.
}
\description{
This is a function for reading a csv type document. External packages are loaded
including readr (for reading a file) and dplyr (for transforming into a tibble data
framework). The file name need to include a directory of the file. Otherwise, such a
function cannot read it successfully.
}
\examples{
fars_read("Desktop/accident_2013.csv")
}
