% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_data.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read months and years from documents.}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{a vector of years.}
}
\value{
a list of tibbles including months, years and errors if exist.
}
\description{
This function is used for read the months and years information from a document.
External package dplyr is loaded to extract month and year information.
}
\examples{
fars_read_years(c(2013,2014))

}
