% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_summarize_years.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize fars record by year}
\usage{
fars_summarize_years(years, path = ".")
}
\arguments{
\item{years}{(num) An integer vector of years (or an object
cohercible to it) containing the years of fars files
of interest.}

\item{path}{(chr) path to the fars file}
}
\value{
A data frame with a number of column equals to the
        number of (valid) years passed as input (plus one for the
        months) and one row each month included in the corresponding
        fars files. The data frame is filled with the number
        of records included for the corresponding year and month.
}
\description{
Compute the number of records are present for fars for each year
(columns) and month (rows).
}
\note{
If a year not present in the data is provided, a warning is
      returned and no column is created for it.
}
\examples{
\dontrun{
    fars_summarize_years(c(2013, 2014))

    fars_summarize_years(c(2012, 2013, 2014))
}
}
