% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_data.R
\name{make_filename}
\alias{make_filename}
\title{Generate a file name in a form of accident_year.csv.bz2.}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A number of a year.}
}
\value{
a name of the file including the year.
}
\description{
This function is simply used for generating a file name without any external packages required.
}
\examples{
make_filename(1980.1)

}
